[TPE-1] Implement authentifacation (#13)

+++ controllers/category_controller.php
@@ -2,7 +2,7 @@
   require_once './models/category_model.php';
   require_once './views/category_view.php';
   require_once './helpers/controller_helper.php';
-
+  require_once './helpers/auth_helper.php';
   class CategoryController
   {
     private $categoryModel;
@@ -13,38 +13,49 @@
       $this->categoryModel = new CategoryModel();
       $this->categoryView = new CategoryView();
       $this->controllerHelper = new ControllerHelper();
+      $this->authHelper = new authHelper();
     }
 
     public function index()
     {
       $categories = $this->categoryModel->index();
-      $this->categoryView->index($categories);
+      $this->categoryView->index($this->authHelper->getLogged(), $categories);
     }
 
     public function showItems($id)
     {
       $categories = $this->categoryModel->index();
       $items = $this->categoryModel->showItems($id);
-      $this->categoryView->showItems($items, $categories);
+      $this->categoryView->showItems($this->authHelper->getLogged(), $items, $categories);
     }
 
     public function delete($id)
     {
-      $this->categoryModel->delete($id);
-      $this->categoryView->default_view();
+      if ($this->authHelper->isLogged()) 
+      {
+        $this->categoryModel->delete($id);
+        $this->categoryView->defaultView($this->authHelper->getLogged(), $this->categoryModel->index());
+      }
     }
 
     public function create()
     {
-      if ($this->controllerHelper->validateParams($_POST))
-        $this->categoryModel->create($_POST);
-      $this->categoryView->default_view();
+      if ($this->authHelper->isLogged()) 
+      {
+        if ($this->controllerHelper->validateParams($_POST))
+          $this->categoryModel->create($_POST);
+        $this->categoryView->defaultView($this->authHelper->getLogged(), $this->categoryModel->index());
+      }
     }
-
+    
     public function put($id)
     {
-      $this->categoryModel->put($id,$_POST);
-      $this->categoryView->default_view();
+      if ($this->authHelper->isLogged()) 
+      {
+        $this->categoryModel->put($id,$_POST);
+        $this->categoryView->defaultView($this->authHelper->getLogged(), $this->categoryModel->index());
+      }
     }
   }
+
 ?>
